requirements.txt
flask==2.3.3
requests==2.32.5
python-dotenv==1.1.1
flask-cors==6.0.1
gunicorn==23.0.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
pytest-playwright==0.7.1
pytest-mock==3.15.1
